% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtdeaths}
\alias{rtdeaths}
\title{Brazilian road traffic deaths data}
\format{
\subsection{\code{rtdeaths} dataset}{

A \code{data.table} with 923,440 rows and 18 columns:
\describe{
\item{\code{cod_modal}}{ICD-10 code of main cause of death}
\item{\code{modal_vitima}}{Mode of transport of the victim during the road crash}
\item{\code{data_ocorrencia}}{Date of death in ISO 8601 format}
\item{\code{ano_ocorrencia}}{Year of death}
\item{\code{idade_vitima}}{Age in years of the victim}
\item{\code{faixa_etaria_vitima}}{Age range of the victim}
\item{\code{sexo_vitima}}{Sex of the victim}
\item{\code{escolaridade_vitima}}{Education of the victim}
\item{\code{raca_vitima}}{Race of the victim}
\item{\code{ocup_cbo_vitima}}{Ocupation of the victim, based on the CBO code}
\item{\code{cod_municipio_ocor}}{City of death occurrence, based on IBGE code}
\item{\code{nome_regiao_ocor}}{Brazilian region where the death occurred}
\item{\code{cod_municipio_res}}{City of residence of the victim, based on IBGE code}
\item{\code{nome_regiao_res}}{Brazilian region where the victim had an residence address}
\item{\code{nome_uf_ocor}}{Name of the state where the death occurred}
\item{\code{nome_municipio_ocor}}{Name of the city where the death occurred}
\item{\code{nome_uf_res}}{Name of the state where the victim had an residence address}
\item{\code{nome_municipio_res}}{Name of the city where the victim had an residence address}
}
}
}
\source{
\url{ftp://ftp.datasus.gov.br/dissemin/publicos/SIM/CID10/DOFET}
}
\usage{
rtdeaths
}
\description{
Death certificates from road traffic crashes. Data from the Brazilian
Ministry of Health, between 1996 and 2021. All variables are encoded in
Brazilian Portuguese.
}
\keyword{datasets}
